+++ src/Collections/LazyDtoCollection.php
@@ -2,10 +2,13 @@
 
 namespace Bfg\Dto\Collections;
 
+use Bfg\Dto\Dto;
 use Bfg\Dto\Traits\DtoCollectionMethodsTrait;
 use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\LazyCollection;
 
-class DtoCollection extends Collection
+class LazyDtoCollection extends LazyCollection
 {
     use DtoCollectionMethodsTrait;
 }