+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.2](https://github.com/bfg-s/dto/compare/1.1.1...1.1.2) - 2024-12-20
+* Added `fromFile` constructor to create a `Dto` from file
+* Added `toApi` method to send data for `POST` request
+* Added `default` method to set default value for property
+* Added deep model support
+
 ## [1.1.1](https://github.com/bfg-s/dto/compare/1.1.0...1.1.1) - 2024-12-09
 * Fixed `make:dto-docs` command, now it generates documentation and for `lazy` properties and for computed properties
 * Fixed reflexion `fromArray...` and `toArray...` now is not computed properties